Refactor cache keys into its own module
